+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 2.3.0 - 2023-01-25
+
+- Fix inclusion of checks when no related project exists.
+
 ## 2.2.2 - 2023-01-16
 
 - Fixed off-chain checks.